Feature/authorization

+++ src/main/java/com/danielabgaryan/mediamatch/service/CommentService.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import com.danielabgaryan.mediamatch.repository.CommentRepository;
 import com.danielabgaryan.mediamatch.repository.MediaListRepository;
 import com.danielabgaryan.mediamatch.repository.UserRepository;
+import com.danielabgaryan.mediamatch.exception.ForbiddenException;
 import com.danielabgaryan.mediamatch.exception.ResourceNotFoundException;
 import com.danielabgaryan.mediamatch.model.Comment;
 import com.danielabgaryan.mediamatch.model.User;
@@ -22,16 +23,18 @@ public class CommentService {
         this.mediaListRepository = mediaListRepository;
     }
 
-    public Comment createComment(Long userId, Long mediaListId, String content) {
-        User user = getUserOrThrow(userId);
-        MediaList mediaList = getMediaListOrThrow(mediaListId);
+    public Comment createComment(String email, Long mediaListId, String content) {
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
+            new ResourceNotFoundException("User not found"));
         
+        MediaList mediaList = getMediaListOrThrow(mediaListId);
         Comment comment = new Comment(user, mediaList, content);
         return commentRepository.save(comment);
     }
 
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, String email) {
         Comment comment = getCommentById(commentId);
+        verifyOwnership(comment, email);
         commentRepository.delete(comment);
     }
 
@@ -51,11 +54,17 @@ public class CommentService {
 
     //helper method to get User or throw a ResourceNotFoundException
     private User getUserOrThrow(Long userId) {
-        return userRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("User Not found"));
+        return userRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("User not found"));
     }
 
     //helper method to get Media List or throw a ResourceNotFoundException
     private MediaList getMediaListOrThrow(Long mediaListId) {
-        return mediaListRepository.findById(mediaListId).orElseThrow(() -> new ResourceNotFoundException("Media List not found"));
+        return mediaListRepository.findById(mediaListId).orElseThrow(() -> new ResourceNotFoundException("Media list not found"));
+    }
+
+    private void verifyOwnership(Comment comment, String email) {
+        if (!comment.getUser().getEmail().equals(email)) {
+            throw new ForbiddenException("You do not own this comment");
+        }
     }
 }
